Start refactor to use smaller components
